sep event func -> file
